agents/simple/pyproject.toml
[project]
name = "simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects",
    "azure-identity>=1.25.1",
    "click>=8.3.0",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
azure-ai-projects = { git = "ssh://git@github.com/azure/azure-sdk-for-python.git", subdirectory = "sdk/ai/azure-ai-projects", rev = "feature/azure-ai-projects/2.0.0b1" }
